Automatic cucumber tests doesn't run on Play 1.2.5
The Play framework has changed the text that is written when it is ready.
Because of this it is not possible to run tests from command line. This fix
checks for a text "@cukes" that are written by the plugin itself and should
work for all versions of Play that the plugin is ment for.